requirements.txt
nltk
pandas
scikit-learn
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
